refactor: 상품 하나에 대한 쿠폰 적용 테스트 수정, 예외처리 수정
