+++ backend/main-service/src/test/java/kkakka/mainservice/order/acceptance/OrderAcceptanceTest.java
@@ -232,65 +232,4 @@ public class OrderAcceptanceTest extends DocumentConfiguration {
 
         return response.body().jsonPath().get("accessToken");
     }
-
-    private String 쿠폰_생성함(String grade, Long productId, String priceRule, Integer percentage,
-        Integer maxDiscount) {
-        if (grade != null) {
-            grade = "\"" + grade + "\"";
-        }
-        final ExtractableResponse<Response> response = RestAssured
-            .given().log().all()
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .body("{\n"
-                + "  \"categoryId\": null,\n"
-                + "  \"grade\": " + grade + ",\n"
-                + "  \"productId\": " + productId + ",\n"
-                + "  \"name\": \"test\",\n"
-                + "  \"priceRule\": \"" + priceRule + "\",\n"
-                + "  \"startedAt\": \"2020-01-01 00:00:00\",\n"
-                + "  \"expiredAt\": \"2025-01-01 00:00:00\",\n"
-                + "  \"percentage\": " + percentage + ",\n"
-                + "  \"maxDiscount\": " + maxDiscount + ",\n"
-                + "  \"minOrderPrice\": 20000\n"
-                + "}")
-            .when()
-            .post("/api/coupons")
-            .then().log().all().extract();
-        return response.header("Location");
-    }
-
-    private String 상품_쿠폰_다운로드(String accessToken) {
-        String couponId = 쿠폰_생성함(null, PRODUCT_1.getId(), "COUPON", 20, 2000);
-
-        final ExtractableResponse<Response> response = RestAssured
-            .given().log().all()
-            .header("Authorization", "Bearer " + accessToken)
-            .when()
-            .post("/api/coupons/download/" + couponId)
-            .then().log().all().extract();
-
-        return couponId;
-    }
-
-    @DisplayName("상품 바로주문 쿠폰 적용 - 성공")
-    @Test
-    void applyProductCoupon_success() {
-        // given
-        String accessToken = 액세스_토큰_가져옴();
-        String couponId = 상품_쿠폰_다운로드(accessToken);
-        ProductOrderDto productOrderDto = new ProductOrderDto(PRODUCT_1.getId(), 3);
-
-        // when
-        final ExtractableResponse<Response> response = RestAssured
-            .given().log().all()
-            .header("Authorization", "Bearer " + accessToken)
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .body(productOrderDto)
-            .when()
-            .post("/api/orders/" + couponId)
-            .then().log().all().extract();
-
-        // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
 }
